requirements.txt
flask
python-dotenv
tqdm
pinecone-client
openai
langchain
langchain-pinecone
langchain-openai
flask_cors
